OneDrive/Desktop/CollabHub/backend/requirements.txt
# Django Core
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.14.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0

# CORS Headers
django-cors-headers>=4.3.0

# Environment variables
python-dotenv>=1.0.0

# Image processing (for avatars)
Pillow>=10.0.0

# Django Filters for API filtering
django-filter>=23.5

# Real-time WebSockets (Phase 4)
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
